tenant TiDB auto-embedding schema contract and starter validation
