+++ pkg/tenant/schema_zero.go
@@ -1,129 +1,11 @@
 package tenant
 
-import (
-	"database/sql"
-	"fmt"
-
-	_ "github.com/go-sql-driver/mysql"
-)
-
-// InitTiDBTenantSchema initializes the tenant data-plane schema on a TiDB/MySQL
-// DSN using the current TiDB-oriented schema shape required by dat9.
+// InitTiDBTenantSchema initializes the TiDB launch schema baseline with the
+// shared database-managed auto-embedding contract used by TiDB tenants.
 func InitTiDBTenantSchema(dsn string) error {
 	return initZeroSchema(dsn)
 }
 
 func initZeroSchema(dsn string) error {
-	if hasMultiStatements(dsn) {
-		return fmt.Errorf("multiStatements is not allowed")
-	}
-	db, err := sql.Open("mysql", dsn)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = db.Close() }()
-	if err := db.Ping(); err != nil {
-		return err
-	}
-	if !isTiDBCluster(db) {
-		return fmt.Errorf("provider requires TiDB capabilities (FTS/VECTOR)")
-	}
-
-	stmts := []string{
-		`CREATE TABLE IF NOT EXISTS file_nodes (
-			node_id      VARCHAR(64) PRIMARY KEY,
-			path         VARCHAR(512) NOT NULL,
-			parent_path  VARCHAR(512) NOT NULL,
-			name         VARCHAR(255) NOT NULL,
-			is_directory BOOLEAN NOT NULL DEFAULT FALSE,
-			file_id      VARCHAR(64),
-			created_at   DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3)
-		)`,
-		`CREATE UNIQUE INDEX idx_path ON file_nodes(path)`,
-		`CREATE INDEX idx_parent ON file_nodes(parent_path)`,
-		`CREATE INDEX idx_file_id ON file_nodes(file_id)`,
-		// See docs/async-embedding/async-embedding-generation-proposal.md,
-		// section "2) File schema: embedding must become mutable and revision-aware".
-		`CREATE TABLE IF NOT EXISTS files (
-			file_id            VARCHAR(64) PRIMARY KEY,
-			storage_type       VARCHAR(32) NOT NULL,
-			storage_ref        TEXT NOT NULL,
-			content_blob       LONGBLOB,
-			content_type       VARCHAR(255),
-			size_bytes         BIGINT NOT NULL DEFAULT 0,
-			checksum_sha256    VARCHAR(128),
-			revision           BIGINT NOT NULL DEFAULT 1,
-			status             VARCHAR(32) NOT NULL DEFAULT 'PENDING',
-			source_id          VARCHAR(255),
-			content_text       LONGTEXT,
-			embedding          VECTOR(1024),
-			embedding_revision BIGINT,
-			created_at         DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3),
-			confirmed_at       DATETIME(3),
-			expires_at         DATETIME(3)
-		)`,
-		`CREATE INDEX idx_status ON files(status, created_at)`,
-		`ALTER TABLE files
-			ADD FULLTEXT INDEX idx_fts_content(content_text)
-			WITH PARSER MULTILINGUAL
-			ADD_COLUMNAR_REPLICA_ON_DEMAND`,
-		`ALTER TABLE files
-			ADD VECTOR INDEX idx_files_cosine((VEC_COSINE_DISTANCE(embedding)))
-			ADD_COLUMNAR_REPLICA_ON_DEMAND`,
-		`CREATE TABLE IF NOT EXISTS file_tags (
-			file_id   VARCHAR(64) NOT NULL,
-			tag_key   VARCHAR(255) NOT NULL,
-			tag_value VARCHAR(255) NOT NULL DEFAULT '',
-			PRIMARY KEY (file_id, tag_key)
-		)`,
-		`CREATE INDEX idx_kv ON file_tags(tag_key, tag_value)`,
-		`CREATE TABLE IF NOT EXISTS uploads (
-			upload_id          VARCHAR(64) PRIMARY KEY,
-			file_id            VARCHAR(64) NOT NULL,
-			target_path        VARCHAR(512) NOT NULL,
-			s3_upload_id       VARCHAR(255) NOT NULL,
-			s3_key             VARCHAR(2048) NOT NULL,
-			total_size         BIGINT NOT NULL,
-			part_size          BIGINT NOT NULL,
-			parts_total        INT NOT NULL,
-			status             VARCHAR(32) NOT NULL DEFAULT 'UPLOADING',
-			fingerprint_sha256 VARCHAR(128),
-			idempotency_key    VARCHAR(255),
-			created_at         DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3),
-			updated_at         DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3) ON UPDATE CURRENT_TIMESTAMP(3),
-			expires_at         DATETIME(3) NOT NULL,
-			active_target_path VARCHAR(512) AS (CASE WHEN status = 'UPLOADING' THEN target_path ELSE NULL END) STORED
-		)`,
-		`CREATE INDEX idx_upload_path ON uploads(target_path, status)`,
-		`CREATE UNIQUE INDEX idx_idempotency ON uploads(idempotency_key)`,
-		// semantic_tasks groups fields by responsibility:
-		// - identity/resource binding: task_id, task_type, resource_id, resource_version
-		// - delivery state: status, attempt_count, max_attempts
-		// - lease/claim ownership: receipt, leased_at, lease_until, available_at
-		// - diagnostics/audit: payload_json, last_error, created_at, updated_at, completed_at
-		// payload_json is only for lightweight hints/debugging; worker correctness
-		// must always re-read current file state via resource_id + resource_version.
-		`CREATE TABLE IF NOT EXISTS semantic_tasks (
-			task_id           VARCHAR(64) PRIMARY KEY,
-			task_type         VARCHAR(32) NOT NULL,
-			resource_id       VARCHAR(64) NOT NULL,
-			resource_version  BIGINT NOT NULL,
-			status            VARCHAR(20) NOT NULL,
-			attempt_count     INT NOT NULL DEFAULT 0,
-			max_attempts      INT NOT NULL DEFAULT 5,
-			receipt           VARCHAR(128) NULL,
-			leased_at         DATETIME(3) NULL,
-			lease_until       DATETIME(3) NULL,
-			available_at      DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3),
-			payload_json      JSON NULL,
-			last_error        TEXT NULL,
-			created_at        DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3),
-			updated_at        DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3),
-			completed_at      DATETIME(3) NULL
-		)`,
-		`CREATE UNIQUE INDEX uk_task_resource_version ON semantic_tasks(task_type, resource_id, resource_version)`,
-		`CREATE INDEX idx_task_claim ON semantic_tasks(status, available_at, lease_until, created_at)`,
-	}
-
-	return execSchemaStatements(db, stmts)
+	return initTiDBAutoEmbeddingSchema(dsn, true)
 }
