+++ pkg/tenant/starter_test.go
@@ -2,12 +2,17 @@ package tenant
 
 import (
 	"context"
+	"strings"
 	"testing"
 )
 
-func TestStarterProvisionerInitSchemaIsNoop(t *testing.T) {
+func TestStarterProvisionerInitSchemaValidatesSchema(t *testing.T) {
 	p := &StarterProvisioner{}
-	if err := p.InitSchema(context.Background(), "ignored-dsn"); err != nil {
-		t.Fatalf("InitSchema should be a no-op for starter: %v", err)
+	err := p.InitSchema(context.Background(), "ignored-dsn")
+	if err == nil {
+		t.Fatal("expected starter schema validation to reject invalid dsn")
+	}
+	if !strings.Contains(strings.ToLower(err.Error()), "unknown network") && !strings.Contains(strings.ToLower(err.Error()), "missing the slash separating the database name") {
+		t.Fatalf("unexpected starter schema validation error: %v", err)
 	}
 }
